backend/requirements.txt
# SportSync API - Python Dependencies
fastapi==0.120.3
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
pydantic==2.7.1
pydantic-settings==2.3.0
PyJWT[crypto]==2.12.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.22
httpx==0.27.0
requests==2.32.5
boto3==1.39.11
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.0
joblib==1.4.2
pytest==8.2.1
pytest-asyncio==0.23.7
